:hammer: Change label input to VectorXi

+++ experiment/experiment/io.h
@@ -7,6 +7,6 @@ const std::string kTrainFilePath = "D:/Andie/code/data-mining-course-project/hom
 const std::string kTestFilePath = "D:/Andie/code/data-mining-course-project/homework2/dataset/t.txt";
 const std::string kTarget = "D:/Andie/code/data-mining-course-project/experiment/output.csv";
 
-std::pair<Eigen::MatrixXd, Eigen::VectorXd> TrainData();
+std::pair<Eigen::MatrixXd, Eigen::VectorXi> TrainData();
 Eigen::MatrixXd TestData();
 void SavePrediction(const Eigen::VectorXd & prediction);
